+++ packages/humanitarian-icons-react/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 11.17.9 (2024-01-23)
+
+**Note:** Version bump only for package @wfp/humanitarian-icons-react
+
+
+
+
+
+## 11.17.8 (2024-01-22)
+
+**Note:** Version bump only for package @wfp/humanitarian-icons-react
+
+
+
+
+
+## 11.17.7 (2024-01-19)
+
+**Note:** Version bump only for package @wfp/humanitarian-icons-react
+
+
+
+
+
 ## 11.17.6 (2024-01-08)
 
 **Note:** Version bump only for package @wfp/humanitarian-icons-react