+++ packages/logos/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 0.3.9 (2024-01-23)
+
+**Note:** Version bump only for package @un/logos
+
+
+
+
+
+## 0.3.8 (2024-01-22)
+
+**Note:** Version bump only for package @un/logos
+
+
+
+
+
+## 0.3.7 (2024-01-19)
+
+**Note:** Version bump only for package @un/logos
+
+
+
+
+
 ## 0.3.6 (2024-01-08)
 
 **Note:** Version bump only for package @un/logos